% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_conversion_milliers.R
\name{conversion_milliers}
\alias{conversion_milliers}
\title{Conversion en milliers d'euros}
\usage{
conversion_milliers(data)
}
\arguments{
\item{data}{le jeu de données}
}
\value{
dataframe : le jeu de données avec les
         données converties en milliers d'euros
}
\description{
Conversion en milliers d'euros
}
\examples{
conversion_milliers(data)
}
