% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_normalize.R
\name{data_normalize}
\alias{data_normalize}
\title{Normalize numerical data using the explotations' surface for farming kind
and the ugb for breeding ones.}
\usage{
data_normalize(data)
}
\arguments{
\item{data}{A data.frame: the full dataset}
}
\value{
A data.frame normalized
}
\description{
Normalize numerical data using the explotations' surface for farming kind
and the ugb for breeding ones.
}
