% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_numeric_only.R
\name{data_numeric_only}
\alias{data_numeric_only}
\title{Restricts data.frame variables to numerical only}
\usage{
data_numeric_only(data)
}
\arguments{
\item{data}{A data.frame: the full dataset}
}
\value{
The data.frame with only it's numerical columns
}
\description{
Restricts data.frame variables to numerical only
}
