% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep_all.R
\name{data_prep_all}
\alias{data_prep_all}
\title{Prepare the dataset: convert to thousands when needed, restrict to
individuals present at least n_min_years, remove outliers and split in
train and test sets}
\usage{
data_prep_all(
  data,
  conv_mil = FALSE,
  n_min_years = 5,
  outliers_custom_cutoff = NULL,
  remove_non_num = TRUE,
  split_pct_train = 0.9
)
}
\arguments{
\item{data}{A data.frame: the full dataset}

\item{conv_mil}{A logical indicating if thousands conversion should be done}

\item{n_min_years}{An integer giving the minimal number of years a farm
should appeared in the data to be selected}

\item{outliers_custom_cutoff}{Threshold to identified outliers}

\item{remove_non_num}{A logicial indicating if non numeric columns should
be ignored}

\item{split_pct_train}{Ratio between train and test data}
}
\value{
A list of data.frame: train and test sets
}
\description{
Prepare the dataset: convert to thousands when needed, restrict to
individuals present at least n_min_years, remove outliers and split in
train and test sets
}
