% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_div_jeu.R
\name{div_jeu}
\alias{div_jeu}
\title{Divise le jeu de données en jeu d'apprentissage et en jeu de test}
\usage{
div_jeu(data, pct_train = 0.9)
}
\arguments{
\item{data}{le jeu de données}

\item{pct_train}{ratio pour le jeu d'apprentissage}
}
\value{
liste : jeu d'apprentissage , jeu de test
}
\description{
Divise le jeu de données en jeu d'apprentissage et en jeu de test
}
\examples{
div_jeu(data)
div_jeu(data, 0.7)
}
