% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{error}
\alias{error}
\title{Returns the error between actual and predicted using metric}
\usage{
error(actual, predicted, metric = "rmse")
}
\arguments{
\item{actual}{A numeric vector of true values}

\item{predicted}{A numeric vector of estimated values}

\item{metric}{A character naming how to compute the error}
}
\value{
The numeric error
}
\description{
Returns the error between actual and predicted using metric
}
\examples{
error(actual = c(35, 36), predicted = c(36, 34), metric = "rmse")
}
