% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{est_percent_below}
\alias{est_percent_below}
\title{Computes the percentage of estimates' errors below a given threshold}
\usage{
est_percent_below(
  actual,
  predicted,
  cutoff = 5,
  metric = "mape",
  errs = ind_error(actual, predicted, metric)
)
}
\arguments{
\item{actual}{A numeric vector of true values}

\item{predicted}{A numeric vector of estimated values}

\item{cutoff}{A numeric: "acceptable" error in percent}

\item{metric}{A character naming how to compute the error}
}
\value{
The numeric percentage of prediction errors below the cutoff
}
\description{
Computes the percentage of estimates' errors below a given threshold
}
\examples{
est_percent_below(actual = c(35, 36), predicted = c(36, 34), cutoff = 5,
                  metric = "mape")
}
