% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ajust.R
\name{get_metric_fun}
\alias{get_metric_fun}
\title{Helper function returning the desired metric function}
\usage{
get_metric_fun(metric)
}
\arguments{
\item{metric}{A character naming the expected metric function}
}
\value{
The function corresponding the character or NULL
}
\description{
Helper function returning the desired metric function
}
\examples{
get_metric_fun("mae")
get_metric_fun("rmse")
get_metric_fun("mape")
}
