% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ajust.R
\name{mae}
\alias{mae}
\title{Calcul de l'écart moyen absolu (MAE)}
\usage{
mae(actual, predicted)
}
\arguments{
\item{actual}{le vecteur des "vraies" valeurs}

\item{predicted}{le vecteur des valeurs estimées}
}
\value{
numeric : l'écart moyen absolu (MAE)
}
\description{
Calcul de l'écart moyen absolu (MAE)
}
\examples{
mae(actual = c(35, 36), predicted = c(36, 34))
}
