% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ajust.R
\name{mape}
\alias{mape}
\title{Calcul de l'écart moyen absolu en pourcentage (MAPE)}
\usage{
mape(actual, predicted)
}
\arguments{
\item{actual}{le vecteur des "vraies" valeurs}

\item{predicted}{le vecteur des valeurs estimées}
}
\value{
numeric : l'écart moyen absolu en pourcentage (MAPE)
}
\description{
Calcul de l'écart moyen absolu en pourcentage (MAPE)
}
\examples{
mape(actual = c(35, 36), predicted = c(36, 34))
}
