% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_backward.R
\name{mod_backward}
\alias{mod_backward}
\title{Backward regression based on adjusted R2}
\usage{
mod_backward(data, x_names, y_name, r2_threshold = 0.01)
}
\arguments{
\item{data}{training dataset}

\item{x_names}{X variables}

\item{y_name}{y variable}
}
\value{
best (ie parsimonious) model
}
\description{
Backward regression based on adjusted R2
}
