% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compress.R
\name{mod_compress}
\alias{mod_compress}
\title{Compress a model in size by removing attributes not needed for prediction}
\usage{
mod_compress(model)
}
\arguments{
\item{model}{Model to reduce}
}
\value{
A compress model
}
\description{
Compress a model in size by removing attributes not needed for prediction
}
\examples{
mod_compress(model)
}
