% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_cv_penalized.R
\name{mod_cv_penalized}
\alias{mod_cv_penalized}
\title{Fits a GLM penalized model using k-fold cross-validation to tune the
alpha and lambda hyperparameters}
\usage{
mod_cv_penalized(
  data,
  X_names = character(),
  y_name,
  alpha = seq(0, 1, by = 0.2),
  cv_strat = "min",
  n_folds = 10,
  type_measure = "mse"
)
}
\arguments{
\item{data}{A numeric dataset in matrix-convertible format}

\item{X_names}{A character vector of explainatory variables names}

\item{y_name}{A character naming the variable to explain}

\item{alpha}{A numeric vector of alpha to use for the cross-validation}

\item{cv_strat}{A character telling which alpha should be consider optimal
"min": minimize the error
"1se":  selects less variables while allowing a greater
        error (min + 1se)}

\item{n_folds}{An integer giving the number of folds tu use in k-folds cv}

\item{type_measure}{A character naming the error metric used for the
cross-validation (defaults to "mse")}
}
\value{
The fitted model
}
\description{
Fits a GLM penalized model using k-fold cross-validation to tune the
alpha and lambda hyperparameters
}
