% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineaire.R
\name{mod_lineaire}
\alias{mod_lineaire}
\title{Entraine et retourne un modèle linéaire}
\usage{
mod_lineaire(data, x_names, y_name)
}
\arguments{
\item{data}{le jeu de données}

\item{x_names}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}
}
\value{
le modèle entrainé
}
\description{
Entraine et retourne un modèle linéaire
}
\examples{
mod_lineaire(data, "RESCO", c("MACHINE.IND", "MILEX"))
}
