% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineaire.R
\name{mod_lineaires}
\alias{mod_lineaires}
\title{Applique des modèles linéaires simples et en retourne un résumé}
\usage{
mod_lineaires(
  data_train,
  data_test,
  X_names,
  y_name,
  metric = "rmse",
  below_cutoff = 5
)
}
\arguments{
\item{X_names}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}

\item{metric}{seuil à dépasser}

\item{data}{le jeu de données}
}
\value{
data.frame (r2 et erreurs)
}
\description{
Applique des modèles linéaires simples et en retourne un résumé
}
\examples{
mod_lineaires(data, "RESCO", c("MACHINE.IND", "MILEX"))
}
