% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineaire.R
\name{mod_lineaires_best_var}
\alias{mod_lineaires_best_var}
\title{Applique des modèles linéaires simples et renvoie le nom des variables
explicatives dont le modèle donne un R² supérieur à un seuil (défaut 70%)}
\usage{
mod_lineaires_best_var(data, x_names, y_name, cutoff = 0.7)
}
\arguments{
\item{data}{le jeu de données}

\item{x_names}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}

\item{cutoff}{seuil à dépasser}
}
\value{
vecteur :    nom des variables explicatives pour lesquelles le modèle
                     linéaire simple renvoie un R² supérieur à cutoff
}
\description{
Applique des modèles linéaires simples et renvoie le nom des variables
explicatives dont le modèle donne un R² supérieur à un seuil (défaut 70%)
}
\examples{
mod_lineaires_best_var(data, "RESCO", c("MACHINE.IND", "MILEX"))
}
