% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mixtes.R
\name{mod_mixtes}
\alias{mod_mixtes}
\title{Applique un modèle mixte}
\usage{
mod_mixtes(
  data,
  y_name,
  effets,
  vars = setdiff(names(data), y_name),
  formule = "auto"
)
}
\arguments{
\item{data}{jeu d'apprentissage du modèle}

\item{y_name}{variable à expliquer}

\item{effets}{effets aléatoires}

\item{vars}{vecteur des noms des variables explicatives
(par défaut toutes les variables sauf y)}

\item{formule}{formule de l'équation du modèle
(par défaut : y_name ~ vars_1 + vars_2 + ... + effets)}
}
\value{
model
}
\description{
Applique un modèle mixte
}
\examples{
mod_mixtes(data_train, "RESCO", "(MILEX | IDNUM)")
}
