% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_performance.R
\name{mod_performance}
\alias{mod_performance}
\title{Returns a basic summary of a model's performance}
\usage{
mod_performance(model, data_test, y_name, metric, below_cutoff)
}
\arguments{
\item{model}{The model to evaluate}

\item{data_test}{A data.frame of test data}

\item{y_name}{The variable to explain}

\item{metric}{The metric to use for error computation}

\item{below_cutoff}{seuil à dépasser}
}
\value{
A data.frame holding the r.squared and errors metrics
}
\description{
Returns a basic summary of a model's performance
}
\examples{
mod_performance(model, data_test, "RESCO", c("MACHINE.IND", "MILEX"))
}
