% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineaire.R
\name{mod_uni_lineaires}
\alias{mod_uni_lineaires}
\title{Entraine des modèles linéaires simples et les retourne}
\usage{
mod_uni_lineaires(data_train, x_names, y_name)
}
\arguments{
\item{data_train}{le jeu de données d'entrainement}

\item{x_names}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}
}
\value{
liste de modèles
}
\description{
Entraine des modèles linéaires simples et les retourne
}
\examples{
mod_uni_lineaires(data_train, "RESCO", c("MACHINE.IND", "MILEX"))
}
