% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{model_error}
\alias{model_error}
\title{Returns the error introduced by a model}
\usage{
model_error(model, data_test, y_name, metric = "rmse")
}
\arguments{
\item{model}{A model}

\item{data_test}{A data.frame}

\item{y_name}{A character naming the variable to explain}

\item{metric}{A character naming how to compute the error}
}
\value{
The numeric error
}
\description{
Returns the error introduced by a model
}
\examples{
model_error(mod_lm, data_test, y_name = "SUBEX", metric = "rmse")
}
