% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_performance.R
\name{model_performance}
\alias{model_performance}
\title{Returns a basic summary of a model's performance}
\usage{
model_performance(
  model,
  data_test,
  y_name,
  metric,
  est_below_cutoff,
  err_above_cutoff
)
}
\arguments{
\item{model}{The model to evaluate}

\item{data_test}{A data.frame of test data}

\item{y_name}{The variable to explain}

\item{metric}{The metric to use for error computation}

\item{est_below_cutoff}{"Acceptable" error in percent}

\item{err_above_cutoff}{Desired percent of "acceptable" estimated values}
}
\value{
A data.frame holding the r.squared and errors metrics
}
\description{
Returns a basic summary of a model's performance
}
\examples{
model_performance(model, data_test, "RESCO")
}
