% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_variables.R
\name{model_variables.glmnet}
\alias{model_variables.glmnet}
\title{Returns the variables' names of non zero betas of an EN fitted model}
\usage{
\method{model_variables}{glmnet}(model, lambda = NULL)
}
\arguments{
\item{lambda}{A numeric corresponding to the desired regularization index}

\item{mod_en}{A fitted EN model}
}
\value{
A character vector of selected variables' names
}
\description{
Returns the variables' names of non zero betas of an EN fitted model
}
