% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_variables.R
\name{model_variables.lm}
\alias{model_variables.lm}
\title{Returns the variables' names of a LM fitted model}
\usage{
\method{model_variables}{lm}(model, lambda = NULL)
}
\arguments{
\item{mod_en}{A fitted LM model}
}
\value{
A character vector of variables' names
}
\description{
Returns the variables' names of a LM fitted model
}
