% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_performance.R
\name{models_performance}
\alias{models_performance}
\title{Returns a basic summary of models' performances}
\usage{
models_performance(
  models,
  names = NULL,
  data_test,
  y_name,
  metric,
  est_below_cutoff,
  err_above_cutoff
)
}
\arguments{
\item{models}{The models to evaluate}

\item{names}{The models' names}

\item{data_test}{A data.frame of test data}

\item{y_name}{The variable to explain}

\item{metric}{The metric to use for error computation}

\item{below_cutoff}{seuil à dépasser}
}
\value{
A data.frame holding the r.squared and errors metrics
}
\description{
Returns a basic summary of models' performances
}
\examples{
models_performance(models, names, data_test, "RESCO",
                   c("MACHINE.IND", "MILEX"))
}
