% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ajust.R
\name{mse}
\alias{mse}
\title{Calcul de l'erreur quadratique moyenne (MSE)}
\usage{
mse(actual, predicted)
}
\arguments{
\item{actual}{le vecteur des "vraies" valeurs}

\item{predicted}{le vecteur des valeurs estimées}
}
\value{
numeric : erreur quadratique moyenne (MSE)
}
\description{
Calcul de l'erreur quadratique moyenne (MSE)
}
\examples{
mse(actual = c(35, 36), predicted = c(36, 34))
}
