% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{percent_err_below}
\alias{percent_err_below}
\title{Computes the percentage of prediction errors below a given threshold}
\usage{
percent_err_below(actual, predicted, cutoff = 5, metric = "mape")
}
\arguments{
\item{actual}{A numeric vector of true values}

\item{predicted}{A numeric vector of estimated values}

\item{cutoff}{A numeric being the threshold value}

\item{metric}{A character naming how to compute the error}
}
\value{
The numeric percentage of prediction errors below the cutoff
}
\description{
Computes the percentage of prediction errors below a given threshold
}
\examples{
percent_err_below(actual = c(35, 36), predicted = c(36, 34), cutoff = 5,
                  metric = "mape")
}
