% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_hat.R
\name{plot_y_hat}
\alias{plot_y_hat}
\title{Graphique y^ en fonction de y}
\usage{
plot_y_hat(model, data_test, y_name)
}
\arguments{
\item{model}{modèle}

\item{data_test}{données à estimer}

\item{y_name}{nom de la variable à expliquer du modèle}
}
\value{
plot : le graphique y^ en fonction de y
}
\description{
Graphique y^ en fonction de y
}
\examples{
plot_y_hat(mod_en, data_test, "SUBEX")
}
