% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_x.R
\name{plot_y_x}
\alias{plot_y_x}
\title{Plot y ~ x}
\usage{
plot_y_x(data, X_names, y_name)
}
\arguments{
\item{data}{le jeu de données}

\item{X_names}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}
}
\value{
les graphiques y ~ x
}
\description{
Plot y ~ x
}
\examples{
plot_y_x(data, "RESCO", c("MACHINE.IND", "MILEX"))
}
