% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_y.R
\name{predict_y}
\alias{predict_y}
\title{Renvoie les y prédits/estimés d'un modèle}
\usage{
predict_y(model, data_test, allow_new_levels = T)
}
\arguments{
\item{model}{modèle de prédiction}

\item{data_test}{jeu de test du modèle}

\item{allow_new_levels}{True par défaut}
}
\value{
vecteur de prédiction de y
}
\description{
Renvoie les y prédits/estimés d'un modèle
}
\examples{
predict_y(mod_en, data_test)
predict_y(mod_lm, data_test)
predict_y(mod_sw, data_test)
predict_y(mod_lmer, data_test)
predict_y(mod_lmer, data_test, allow_new_levels = F)
}
