% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_by.R
\name{reduce_by}
\alias{reduce_by}
\title{Reduces a data.frame numerical variables using another numerical variable}
\usage{
reduce_by(data, reduce_by)
}
\arguments{
\item{data}{A data.frame: a dataset}

\item{reduce_by}{A character naming the variable to reduce by}
}
\value{
The reduced data.frame
}
\description{
Reduces a data.frame numerical variables using another numerical variable
}
