% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_ajust.R
\name{rmse}
\alias{rmse}
\title{Calcul de l'écart quadratique moyen (RMSE)}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{le vecteur des "vraies" valeurs}

\item{predicted}{le vecteur des valeurs estimées}
}
\value{
numeric : l'écart quadratique moyen (RMSE)
}
\description{
Calcul de l'écart quadratique moyen (RMSE)
}
\examples{
rmse(actual = c(35, 36), predicted = c(36, 34))
}
