% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Main function}
\usage{
run(
  data = NULL,
  conv_mil = FALSE,
  n_min_years = 5,
  outliers_custom_cutoff = NULL,
  split_pct_train = 0.9,
  remove_non_num = TRUE,
  data_train = NULL,
  data_test = NULL,
  x_names = NULL,
  x_exclude = NULL,
  y_name,
  metric = "rmse",
  below_cutoff = 5
)
}
\arguments{
\item{data}{A data.frame: the full dataset}

\item{conv_mil}{A logical indicating if thousands conversion should be done}

\item{n_min_years}{An integer giving the minimal number of years a farm
should appeared in the data to be selected}

\item{outliers_custom_cutoff}{Threshold to identified outliers}

\item{split_pct_train}{Ratio between train and test data}

\item{remove_non_num}{A logicial indicating if non numeric columns should
be ignored}

\item{data_test}{A data.frame: the testing dataset}

\item{x_names}{A character vector: explainatory variables}

\item{x_exclude}{A character vector: explainatory variables to exclude}

\item{y_name}{A character: the variable to explain}

\item{metric}{A character: the metric to use for error computation}

\item{below_cutoff}{A numeric: seuil à dépasser}

\item{data_trian}{A data.frame: the training dataset}
}
\value{
A list of both models and their performances
}
\description{
Main function
}
