% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_selection_fermes.R
\name{selection_fermes}
\alias{selection_fermes}
\title{Sélection des fermes présentes au moins X ans (par défaut 5 ans)}
\usage{
selection_fermes(data, nb_annees = 5)
}
\arguments{
\item{data}{le jeu de données (avec la variable IDNUM)}

\item{nb_annees}{le nom de la variable à expliquer}
}
\value{
dataframe : le jeu de données avec les fermes présentes
}
\description{
Sélection des fermes présentes au moins X ans (par défaut 5 ans)
}
\examples{
selection_fermes(data_ind)
}
