% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_vif.R
\name{vif_to_df}
\alias{vif_to_df}
\title{Retourne un data.frame à partir d'un objet retourné par car::vif}
\usage{
vif_to_df(obj, ...)
}
\arguments{
\item{obj}{l'objet retourné par car::vif}
}
\value{
un data.frame donnant le vif de chaque variable
}
\description{
Retourne un data.frame à partir d'un objet retourné par car::vif
}
\examples{
vif_to_df(vif_values)
}
