% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{err_percent_above}
\alias{err_percent_above}
\title{Computes the accepted error in percentage needed to get a percentage of
acceptable estimates close to a target value}
\usage{
err_percent_above(
  actual = NULL,
  predicted = NULL,
  target = 80,
  metric = "mape",
  errs = ind_error(actual, predicted, metric),
  window = list(below = 0, above = 100)
)
}
\arguments{
\item{actual}{A numeric vector of true values}

\item{predicted}{A numeric vector of estimated values}

\item{target}{A numeric: desired percent of "acceptable" estimated values}

\item{metric}{A character naming how to compute the error}

\item{errs}{A vector of numeric being the individual errors}

\item{window}{A list of integers: window where the objective lies}
}
\value{
The numeric percentage of prediction errors below the cutoff
}
\description{
Computes the accepted error in percentage needed to get a percentage of
acceptable estimates close to a target value
}
\examples{
err_percent_above(actual = c(35, 36), predicted = c(36, 34), target = 80,
                  metric = "mape")
}
