% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineaires.R
\name{mod_lineaires}
\alias{mod_lineaires}
\title{Applique des modèles linéaires simples et renvoie le nom des variables explicatives dont le modèle donne un R² supérieur à 70%}
\usage{
mod_lineaires(x_name, y_name, data)
}
\arguments{
\item{x_name}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}

\item{data}{le jeu de données}
}
\value{
vecteur : nom des variables explicatives pour lesquelles le modèle linéaire simple renvoie un R² supérieur à 70%
}
\description{
Applique des modèles linéaires simples et renvoie le nom des variables explicatives dont le modèle donne un R² supérieur à 70%
}
\examples{
mod_lineaires("RESCO", c("MACHINE.IND", "MILEX"), data)
}
