% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_lineaire.R
\name{mod_lineaires}
\alias{mod_lineaires}
\title{Applique des modèles linéaires simples et en retourne un résumé}
\usage{
mod_lineaires(
  data_train,
  data_test,
  x_names,
  y_name,
  metric = "rmse",
  below_cutoff = 5
)
}
\arguments{
\item{data_train}{le jeu de données d'entrainement}

\item{data_test}{le jeu de données de test}

\item{x_names}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}

\item{metric}{metrique d'erreur}

\item{below_cutoff}{seuil à dépasser}
}
\value{
data.frame (r2 et erreurs)
}
\description{
Applique des modèles linéaires simples et en retourne un résumé
}
\examples{
mod_lineaires(data_train, data_test, "RESCO", c("MACHINE.IND", "MILEX"))
}
