% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mixtes.R
\name{mod_mixtes}
\alias{mod_mixtes}
\title{Applique un modèle mixte}
\usage{
mod_mixtes(data, x_names, y_name, effets, formule = "auto")
}
\arguments{
\item{data}{jeu d'apprentissage du modèle}

\item{x_names}{vecteur des noms des variables explicatives}

\item{y_name}{variable à expliquer}

\item{effets}{effets aléatoires}

\item{formule}{formule de l'équation du modèle
(par défaut : y_name ~ vars_1 + vars_2 + ... + effets)}
}
\value{
model
}
\description{
Applique un modèle mixte
}
\examples{
mod_mixtes(data_train, "RESCO", "(MILEX | IDNUM)")
}
