% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_penalized.R
\name{mod_penalized}
\alias{mod_penalized}
\title{Fits a GLM penalized model using k-fold cross-validation to tune the
lambda hyperparameter}
\usage{
mod_penalized(
  data,
  X_names = character(),
  y_names,
  alpha = 1,
  type_measure = "rmse"
)
}
\arguments{
\item{data}{A numeric dataset in matrix-convertible format}

\item{X_names}{A character vector of explainatory variables names}

\item{y_names}{A character naming the variable to explain}

\item{alpha}{A numeric to compromise between ridge (alpha=0) and lasso
(alpha=1, default) penalization}

\item{type_measure}{A character naming the error metric used for the
cross-validation (defaults to "rmse")}
}
\value{
The fitted model
}
\description{
Fits a GLM penalized model using k-fold cross-validation to tune the
lambda hyperparameter
}
