% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stepwise.R
\name{mod_stepwise}
\alias{mod_stepwise}
\title{Régression stepwise (pas-à-pas)}
\usage{
mod_stepwise(x_name, y_name, data, direction = "both")
}
\arguments{
\item{x_name}{les noms des variables explicatives}

\item{y_name}{le nom de la variable à expliquer}

\item{data}{le jeu de données}

\item{direction}{la méthode utilisée dans le stepwise : "backward", "forward" ou "both (par défaut)}
}
\value{
model : le meilleur modèle (basé sur l'AIC) par la méthode stepwise
}
\description{
Régression stepwise (pas-à-pas)
}
\examples{
mod_stepwise(x_name, y_name, data_train, "backward")
}
