% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_y_hat.R
\name{plot_y_hat}
\alias{plot_y_hat}
\title{Graphique y^ en fonction de y}
\usage{
plot_y_hat(actual, predicted)
}
\arguments{
\item{actual}{le vecteur des "vraies" valeurs}

\item{predicted}{le vecteur des valeurs estimées}
}
\value{
plot : le graphique y^ en fonction de y
}
\description{
Graphique y^ en fonction de y
}
\examples{

}
