% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_y.R
\name{predict_y}
\alias{predict_y}
\title{Applique un modèle mixte}
\usage{
predict_y(model, data_test, allow_new_levels = T)
}
\arguments{
\item{model}{modèle de prédiction}

\item{data_test}{jeu de test du modèle}

\item{allow_new_levels}{True par défaut}
}
\value{
vecteur de prédiction de y
}
\description{
Applique un modèle mixte
}
\examples{
mod_mixtes(data_train, "RESCO", "(MILEX | IDNUM)")
}
