% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error.R
\name{rmse}
\alias{rmse}
\title{Calcul de la racine de l'erreur quadratique moyenne (RMSE)}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{le vecteur des "vraies" valeurs}

\item{predicted}{le vecteur des valeurs estimées}
}
\value{
numeric : racine de l'erreur quadratique moyenne (RMSE)
}
\description{
Calcul de la racine de l'erreur quadratique moyenne (RMSE)
}
\examples{
rmse(actual = c(35, 36), predicted = c(36, 34))
}
