% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supp_outliers.R
\name{supp_outliers}
\alias{supp_outliers}
\title{Suppression des outliers}
\usage{
supp_outliers(data, graph = FALSE)
}
\arguments{
\item{data}{le jeu de données}

\item{graph}{pour représenter les points (TRUE ou FALSE)}
}
\value{
dataframe ou plot : le jeu de données sans les outliers ou la représentation graphique
}
\description{
Suppression des outliers
}
\examples{
supp_outliers(data_ind_5y, TRUE)
}
